employee_record: Extract generic helpers into their own modules
