+++ itou/employee_record/typing.py
@@ -1,0 +1,6 @@
+import typing
+
+
+class CodeComInsee(typing.TypedDict):
+    codeComInsee: str | None
+    codeDpt: str
